fix: link against proper libgcc library
Code was being linked against /thumb/nofp instead of the proper /thumb/v7e-m+fp/hard
lib that this chip supports. This introduces the proper functions to
deal with 64 bit integers and doubles (used by printf). For performance
purposes they will not be used anyways, but is good to have them.

This also eliminates all the shenanigans needed to print a floating
number, it can now  be done normally with %f.

The printf library was moved to its develop branch as well to be able to
use floats instead of doubles internally, reducing code size and
probably increasing performance thanks to the FP unit in the chip.