feat: add usb shell echo
This is a basic program to test the implementation of every day
things we take for granted in a normal shell session like seeing
the stuff you type on or backspace actually deleting characters.

It also implements serial specific stuff like waiting for the host to
connect to the serial line before printing out stuff.